arbitrum-client: contract_interaction: set gas price to 2x basefee
